add case auto approve for HC form

+++ eActForm/eActForm/Controllers/ActivityController.cs
@@ -320,7 +320,7 @@ namespace eActForm.Controllers
                                     // case form benefit will auto approve
                                     if (QueryGetBenefit.getAllowAutoApproveForFormHC(activityId))
                                     {
-                                        ApproveAppCode.updateApprove(activityId, AppCode.ApproveStatus.อนุมัติ.ToString(), "", AppCode.ApproveType.Activity_Form.ToString());
+                                        ApproveAppCode.updateApprove(activityId, ((int)AppCode.ApproveStatus.อนุมัติ).ToString(), "", AppCode.ApproveType.Activity_Form.ToString());
                                     }
                                 }
                                 var rtn = await EmailAppCodes.sendApproveAsync(activityId, AppCode.ApproveType.Activity_Form, false);
